uemoa_invoicer_starter/backend/inventory/serializers.py
from rest_framework import serializers
from .models import Supplier, StockMovement

class SupplierSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supplier
        fields = "__all__"

class StockMovementSerializer(serializers.ModelSerializer):
    class Meta:
        model = StockMovement
        fields = "__all__"
