uemoa_invoicer_starter/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3
django-filter>=24.2
celery>=5.3
redis>=5.0
weasyprint>=61.0
psycopg2-binary>=2.9
